requirements.txt
pandas>=1.1.0
tabula-py>=2.1.1
XlsxWriter>=1.3.3